validate auth headers - refresh OAuth access token
